Added better error handling
It's really just a band-aid fix to cover up the fact that we "finish" by
getting back a 304 error from the server and assuming that that means
we're at the end of the chat. Still, at least now we shouldn't
accidentally stop early if we try to retrieve an image that isn't there
anymore.